# docs/swagger.yaml
basePath: /
definitions:
  request.Login:
    properties:
      captcha:
        type: string
      captchaId:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  request.Register:
    properties:
      authorityIds:
        items:
          type: integer
        type: array
      authorityid:
        type: integer
      enable:
        type: integer
      headerImg:
        type: string
      nickname:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  response.LoginResponse:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/system.SysUser'
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  response.SysCaptchaResponse:
    properties:
      captchaId:
        type: string
      captchaLength:
        type: integer
      picPath:
        type: string
    type: object
  response.SysUserResponse:
    properties:
      user:
        $ref: '#/definitions/system.SysUser'
    type: object
  system.SysApi:
    properties:
      apiGroup:
        description: api组
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: api中文描述
        type: string
      id:
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysArticleType:
    properties:
      createdAt:
        description: 创建时间
        type: string
      icon:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysAuthority:
    properties:
      authorityId:
        type: integer
      authorityName:
        type: string
      children:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      default:
        type: string
      deletedAt:
        type: string
      menus:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      parentId:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenu:
    properties:
      authoritys:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      children:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      closeTab:
        type: boolean
      component:
        type: string
      createdAt:
        description: 创建时间
        type: string
      defaultMenu:
        type: boolean
      hidden:
        type: boolean
      icon:
        type: string
      id:
        type: integer
      keepAlive:
        type: boolean
      menuBtn:
        items:
          $ref: '#/definitions/system.SysBaseMenuBtn'
        type: array
      name:
        type: string
      parameters:
        items:
          $ref: '#/definitions/system.SysBaseMenuParameter'
        type: array
      parentId:
        type: string
      path:
        type: string
      sort:
        type: integer
      title:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuBtn:
    properties:
      createdAt:
        description: 创建时间
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      sysBaseMenuId:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuParameter:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        type: integer
      key:
        type: string
      sysBaseMenuID:
        type: integer
      type:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      value:
        type: string
    type: object
  system.SysUser:
    properties:
      authorities:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      authority:
        $ref: '#/definitions/system.SysAuthority'
      authorityId:
        description: 用户角色ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      email:
        type: string
      enable:
        type: integer
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      username:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
  description: Go 语言编程之旅：一起用 Go 做项目
  title: 博客系统 Swagger
  version: 0.0.1
paths:
  /api/createApi:
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 创建基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建基础api
      tags:
      - SysApi
  /article/type/add:
    post:
      parameters:
      - description: 标题，图标
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysArticleType'
      produces:
      - application/json
      responses:
        "200":
          description: 新建标题
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 新增博客一级标题
      tags:
      - ArticleType
  /article/type/delete:
    delete:
      parameters:
      - description: 标题，图标
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysArticleType'
      produces:
      - application/json
      responses:
        "200":
          description: 删除标题
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 删除博客一级标题
      tags:
      - ArticleType
  /article/type/edit:
    put:
      parameters:
      - description: 标题，图标
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysArticleType'
      produces:
      - application/json
      responses:
        "200":
          description: 修改标题
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 修改博客一级标题
      tags:
      - ArticleType
  /article/type/list:
    get:
      parameters:
      - description: 标题，图标
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysArticleType'
      produces:
      - application/json
      responses:
        "200":
          description: 修改标题
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 获取博客一级标题
      tags:
      - ArticleType
  /base/captcha:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 生成验证码,返回包括随机数id,base64,验证码长度
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.SysCaptchaResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 生成验证码
      tags:
      - Base
  /base/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 返回包括用户信息,token,过期时间
          schema:
            allOf:
            - $ref: '#/definitions/response.LoginResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.LoginResponse'
                msg:
                  type: string
              type: object
      summary: 用户登录
      tags:
      - Base
  /user/register:
    post:
      parameters:
      - description: 用户米，昵称，密码，角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: 用户注册账号，返回包括用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.SysUserResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.SysUserResponse'
                msg:
                  type: string
              type: object
      summary: 用户注册账号
      tags:
      - SysUser
swagger: "2.0"